+++ Src/test-fixtures/hakeem-hardening/XPAR.m
@@ -1,0 +1,2 @@
+XPAR ; Parameter APIs
+GET(ENT,PAR) Q 1